+++ packages/provider/src/provider/index.ts
@@ -1,1 +1,2 @@
+export * from './v3/index';
 export * from './v2/index';